Cambios03

+++ Models/equipos.cs
@@ -12,7 +12,7 @@ namespace _2019BB601.Models
         public int marca_id { get; set; }
         public string modelo { get; set; }
         public int anio_compra { get; set; }
-        public double costo { get; set; }
+        public decimal costo { get; set; }
         public int vida_util { get; set; }
         public int estado_equipo_id { get; set; }
         public string estado { get; set; }
